pa-doc-gen/pa-doc-gen/requirements.txt
PySide6>=6.5.0
reportlab>=4.0
PyYAML>=6.0
Markdown>=3.4
Pygments>=2.15
